tools/requirements.txt
google-cloud-bigquery==3.11.4
click